db: Compact Juno database migration
Compact Juno database migrations into a single migration, '254_juno.py'.

Users will now need to update Juno before updating to Kilo or later.

Specific changes include:

- Alter the type of 'user_id' column of 'volume_usage_cache' table from
  'String(32)' to 'String(64)'
- Add 'mtu', 'dhcp_server', and 'share_dhcp' columns to 'networks' table
- Add foreign key constraint from 'compute_node_id' column of
  'pci_devices' table to 'compute_nodes' table
- Alter the 'resource' column of 'quota_usages' table and 'product_id',
  'vendor_id' and 'dev_type' columns of 'pci_devices' table to be
  non-nullable. Alter the 'deleted' column of 'pci_devices' table to be
  nullable
- Drop a duplicate index from the 'block_device_mapping' table
- Drop the 'instance_group_metadata' table
- Add 'instance_extra' table
- Add 'pci_requests' column to 'instance_extra' table
- Add 'request_id' column to 'pci_devices' table

Note that we don't include '248_add_expire_reservations_index' since
that is a duplicate of the old '234_add_expire_reservations_index'
migration which we previously handled in change
I17db7cdaad2c6368092b4fb00d5959711ad249f9.

Diffing the tables using the tool described at [1] highlights a bug in
sqlalchemy-migrate whereby adding a new 'Boolean' column on backends
that do not support a native boolean type does not result in the
addition of the check constraints we see added by sqlalchemy itself.
This is mostly worked around by explicitly disabling the creation of the
check constraint, but due to *another* bug, the constraints are still
added for the shadow table resulting in the following diff (note: many
hunks are excluded since they're the result on non-stable sorting of
output from the '.dump' command):

  $ diff 215-sqlite.sql 253-sqlite.sql
  --- 215-sqlite.sql      2021-01-07 10:06:51.524103889 +0000
  +++ 253-sqlite.sql      2021-01-07 10:40:24.603224423 +0000
  @@ -887,10 +919,16 @@
          uuid VARCHAR(36),
          priority INTEGER,
          rxtx_base INTEGER,
  -       deleted INTEGER, mtu INTEGER, dhcp_server VARCHAR(39), enable_dhcp BOOLEAN, share_address BOOLEAN,
  +       deleted INTEGER,
  +       mtu INTEGER,
  +       dhcp_server VARCHAR(39),
  +       enable_dhcp BOOLEAN,
  +       share_address BOOLEAN,
          PRIMARY KEY (id),
          CHECK (injected IN (0, 1)),
  -       CHECK (multi_host IN (0, 1))
  +       CHECK (multi_host IN (0, 1)),
  +       CHECK (enable_dhcp IN (0, 1)),
  +       CHECK (share_address IN (0, 1))
   );
   CREATE TABLE shadow_pci_devices (
          created_at DATETIME,

I suspect the bug is caused by the copying of the column and we could
probably work around this by special-casing the creation of these
columns in the shadow tables. However, given sqlite is not a production
database this doesn't seem worth the effort. There are no changes in the
diffs of the PostgreSQL and MySQL schemas, aside from a bit of
reshuffling in the latter due to the output of 'mysqldump' not being
stable sorted.

Also note that this exposes quite a few bugs where we either forget to
migrate a shadow table column or, in one case, add something to a shadow
table that we shouldn't have. All of these will have to be resolved in
a later migration, assuming one of the as-yet unsquashed migrations
don't actually address this.

When testing, the previous base version was 233. It is now 253.

[1] https://that.guru/blog/comparing-nova-db-migrations/

Change-Id: I236998bea3619b7cb807023291c31680e8997c40
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>