db: Compact Stein database migrations
Compact Stein database migrations into a single migration,
'390_stein.py'.

Users will now need to update to Stein before updating to Train or
later.

Specific changes include:

- Add 'volume_type' column to 'block_device_mapping' table (391)

Note that this change is unusual in that there are no placeholder
migrations to clean up from the previous release. This is because we
forgot to include them during the Rocky release.

When testing, the previous base version was 390. It is now 391.

Change-Id: Id77311b7b46b8230c0ce23f8eeb8d4dca7b38e88
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>