db: Compact Train database migrations
Compact Train database migrations into a single migration,
'402_train.py'.

Users will now need to update to Train before updating to Ussuri or
later.

Specific changes include:

- Add 'cross_cell_move' column to 'migrations' table (397)
- Add 'vpmems' column to 'instance_extra' table (398)
- Add 'hidden' column to 'instances' table (399)
- Add 'user_id' and 'project_id' columns to 'migrations' table (401)
- Add 'resources' column to 'instance_extra' table (402)

As there are boolean tables here, we're once again hitting the issue
whereby we need to explicitly disable contraints for boolean fields to
keep the schemas the same.

When testing, the previous base version was 391. It is now 402.

Change-Id: Iecb531728a83194117e0e7e5f4b6189073541f68
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>