db: Compact Kilo database migrations
Compacts Kilo database migrations into a single migration,
'280_kilo.py'.

Users will now need to update to Kilo before updating to Liberty of
later.

This change is significantly more complicated than previous compactions,
mostly due to the changes to indexes specifically for PostgreSQL and
SQLite.

Specific changes include:

- Drop duplicate index covering 'host' and 'deleted' field of
  'instances' table (265)
- Drop duplicate index covering 'host' field of 'iscsi_targets' table
  (265)
- Add 'tags' table (266)
- Alter 'uuid' column of 'instances' table to be non-nullable (267)
- Add unique constraint to 'uuid' column of 'instances' table (267)
- Add 'host' column to 'compute_nodes' table (268)
- Add unique constraint to 'host' and 'hypervisor_hostname' columns of
  'compute_nodes' table (268)
- Add 'numa_node' column to 'pci_devices' table (269)
- Add 'flavor' column to 'instance_extra' table (270)
- Add a number of missing indexes to multiple tables on SQLite and
  PostgreSQL (271)
- Add a number of missing foreign keys and unique constraints to
  multiple tables on SQLite (273)
- Add index covering 'project_id' and 'deleted' columns of 'instances'
  table (274)
- Drop index covering 'project_id' column of 'instances' table (274)
- Add 'type' column to 'key_pairs' table (275)
- Add 'vcpu_model' column to 'instance_extra' table (276)
- Add index covering 'deleted', 'allocated', and 'updated_at' columns of
  'fixed_ips' table (277)
- Alter 'service_id' column of 'compute_nodes' table to be nullable
  (278)
- Drop foreign key constraint on 'service_id' column of 'compute_nodes'
  table (278)
- Drop unique constraint covering 'host' and 'hypervisor_hostname'
  columns of 'compute_nodes' table in favour of constraint covering
  'host', 'hypervisor_hostname' and 'deleted' columns (279)
- Alter 'name' column of 'key_pairs' table to be non-nullable (280)

Once again, we have a case where we've forgotten to apply a change to
the shadow table. We also have our first instance of table, the 'tags'
table, that does is not soft deletable and therefore does not have a
shadow table. Both are handled and the former can be resolved later.

Also note that we encounter a small bug in sqlalchemy-migrate along with
a possible bug in sqlalchemy itself. For the former, declaration of an
index before a foreign key that affects the same column will result in
it attempting to drop the index, even if it was never created. For the
latter, attempting to create a table that shares an enum does not work
on PostgreSQL. See comments inline for both issues.

Now that we can rely on the 'instances.uuid' column being unique and
non-nullable, we can remove the 'null_instance_uuid_scan' nova-manage
command. This will be done separately.

When testing, the previous base version was 253. It is now 279.

Change-Id: I2aae01ed235f1257b0b3ddc6aee4efc7be38eb6e
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>