Omit resource inventories from placement update if zero
When a compute node has zero total available for the:

  * MEMORY_MB
  * DISK_GB
  * VGPU
  * PMEM_NAMESPACE_*

resource classes, we attempt to PUT an inventory with 'total' of 0
which isn't allowed by the placement API. Doing this results in a 400
error from placement "JSON does not validate: 0 is less than the
minimum of 1" and ResourceProviderUpdateFailed and
ResourceProviderSyncFailed raised in nova.

We are already omitting most resource classes when their total amount
of the resource is 0 and we just need to also do it for the
aforementioned resource classes.

Closes-Bug: #1901120
Closes-Bug: #1906494

Change-Id: I022f3bbddbbdc24362b10004f273da2421788c97
(cherry picked from commit 0b942dcaa5db81f927d1a300cb839b4e6c4d1c5e)