Add upgrade check about old computes
Report a warning during upgrade checks if there are computes older than
the previous major nova release in the system.

So if code is upgraded to N+1 and the upgrade check was run before
the restart of the services with N+1 code then the check warns for N-1
computes in the system.

There are small modification in the documentation and code comment to
align it to the Victoria context.

Related-Bug: #1903545

Change-Id: I873b0c1e6e695ae88241bbf75ac9f80ecc6f5664
(cherry picked from commit 3b8257cd239d0bb52cb654d709cc579ca9841628)