Bump the Compute RPC API to version 6.0
Even if we want to bump the compute rpc api to a new major version, we still
need to support 5.x for the rolling upgrade case.

Adding a proxy that will handle those old RPC calls/casts so we could easily
cut it in a later change in the next X release timeframe (as we won't longer
support Victoria 5.x calls)

On the client side, we're making sure that we can still support v5.x methods
until the next X release (as operators could pin on 5.0 version for Victoria
compute nodes)

NOTE(sbauza): Now that the compute service only supports the 6.0
version, we could remove all the unnecessary verifications in both the
manager but also the functional and unittests. That being said, given
the patch would be large, I think it's better to just amend the needed
modifications for supporting 6.0 and delete the above in later changes.
For example, the NUMA live migration functests verify a lot of RPC API
compatibities in betwen 5.2, 5.3 and later. We need to stop verifying
this but let's do this in later changes, please.

Change-Id: Id2e467c43d5b7c6bafd5dbbbf8d5d06be6dc7e30