Support Cpu Compararion on Aarch64 Platform
The general use case of libvirt's compareCPU() API is to allow comparing
a guest CPU to a host CPU. However, for AArch64, the _only_ use case
libvirt supports is to allow comparing host CPU with a destination CPU
which is accomplished with the help of this patch[1].

So what we're talking about here, in context of AArch64, is comparing
source _host_ CPU to destination _host_ CPU, so as to prevent a guest
from migrating to a completely different host. For AArch64, in the
past we skipped[2] _guest_ CPU to host CPU comparison (the
"traditional" use case) check for AArch64 for two main reasons, and
these reasons are still valid:
1. There are lots of vendors making different AArch64 CPUs, and they
   are all not easily comparable because they all differ in various
   ways.
2. QEMU AArch64 developers themselves recommend that using 'host-
   passthrough' is  the way to run KVM guests on AArch64.

In sum, for AArch64, (a) guests should use 'host-passthrough';
and (b) src host to dest host CPU comparison only makes sense to
know about _host_ CPU compatibility, but you will not get any CPU
config that you can give to a guest. For more detail discussion,
please check link [3] and discussion [4].

[1] https://listman.redhat.com/archives/libvir-list/2020-September/msg01391.html
    — Modify virCPUarmCompare to perform compare actions
[2] https://review.opendev.org/c/openstack/nova/+/589769/1/nova/virt/libvirt/driver.py#6892
[3] https://www.redhat.com/archives/libvir-list/2020-September/msg00262.html
[4] https://www.redhat.com/archives/libvir-list/2020-September/msg00328.html,

Change-Id: I9026f2ec39660408e4ac66fc0a85223e4b25e4d1
Signed-off-by: Kevin Zhao <kevin.zhao@linaro.org>