[stable-only] Fix lower-constraints job
This patch fixes the newly encountered issues with the
lower-constraints job (came up with the new pip version with the new
dependency resolver) and sets back the job as voting job.

install_command was not yet removed from tox.ini in stable/stein and
seemingly it is used in lower-constraints job, thus the job fails with
contradicting constraints over package 'coverage'. As a minimal fix
the install_command is added to the lower-constraints tox target, too.

warlock==1.3.0 is set as lower-constraint, but during the installation
a requirement to warlock!=1.3.0 causes failure of the lower-constraints
job.

* pyroute2 version bumped to match os-vif's constraint
* psycopg2 and oslo.messaging bumped to make lower-constraints job pass
* zVMCloudConnector supports python 3.6 only from vesion 1.2.2
* cryptography version needed to be bumped as it conflicts with another
  dependency and causes the new pip resolver to time out at gate

Change-Id: Iae3fcac484f060e8dbeef299d594b8ade8ab3b70