Disable qos resize tempest test for nova-multi-cell job
It turned out that during the qos resize work we did not implemented
support of cross cell resize with qos ports. Tempest test coverage for
the resize and migrate is landed recently that made the nova-multi-cell
job to fail.

So this patch disables the qos resize and migrate tempest tests for the
nova-multi-cell job to unblock the gate.

Related-Bug: #1907522

[1] I8c573c4a11a76932057d8415d76114a03269b9f3

Change-Id: I95bc22f7d65454cd9e7b54a0e6d9516f2f204978