[stable-only] fix lower-constraints and disable qos resize
This is a combination of 2 commits, as both are needed to unblock
Ussuri gate.

1. [stable-only] lower-contraints: Bump various packages

* amqp==2.5.0 & kombu==4.6.1 via oslo.messaging
  I4b72d8feb85c2b9b4657510c356cd21e22fe40c2

* enum34==1.1.8 to workaround known issues with the package

  The associated repo and issue tracker is currently offline so I can't
  link the actual issue here but it's widely reported elsewhere.

  https://github.com/iterative/dvc/issues/1995

* hacking==3.0.1 introduced via Ie00c10332bd7110169dbb150d601c157b6694d05

* oslo.context==2.22.0 introduced via oslo-policy 3.1.0

* stestr==2.0.0 introduced via oslotest 3.8.0

2. Disable qos resize tempest test for nova-multi-cell job

It turned out that during the qos resize work we did not implemented
support of cross cell resize with qos ports. Tempest test coverage for
the resize and migrate is landed recently that made the nova-multi-cell
job to fail.

So this patch disables the qos resize and migrate tempest tests for the
nova-multi-cell job to unblock the gate.

Related-Bug: #1907522

[1] I8c573c4a11a76932057d8415d76114a03269b9f3
(cherry picked from commit 048250a4b76c0b6a85e459f92b436c448d592235)

Change-Id: Ib0e09df9d61e42592c2d36ceed7f38c1278c53b2
Closes-Bug: #1907756