libvirt: Record the machine_type of instances in system_metadata
This change starts to record the machine type of instances registered on
a given host during init_host or later in _configure_guest_by_virt_type
when spawning a new instance.

The machine type is recorded in the system metadata of the instance
under the ``image_hw_machine_type`` key as already used to store the
image metadata property ``hw_machine_type``. As a result no changes are
required to the code of libvirt_utils.get_machine_type that looks up the
machine type of an instance based on it's image metadata and host
config.

Functional tests are included to verify the basic behaviour of both this
new registration code and now this leads to instances being effectively
pinned to the machine type used during their initial spawn.

blueprint: libvirt-default-machine-type
Change-Id: I30c780a7729f1e7d791256bdc38d73b976c56268