api: Reject requests to detach a volume when the compute is down
When detaching a volume from an instance Nova will currently always
attempt to make a cast RPC call before returning a successful 202
response to the caller.

The lack of checking here regarding the underlying compute can lead to
requests never being acted upon and has often confused callers when the
compute itself is down.

This change aims to correct this by having the state of the compute
hosting the instance checked in the compute API layer. When the compute
is down the initial request is now rejected with a 409 conflict to the
caller so they understand why the volume still appears attached.

Closes-Bug: #1909120
Change-Id: I759aa36dc00a6c0612b9755dacd9aa414c408498