Refactor ResourceRequest constructor
This refactor changes ResourceRequest __init__ to make only an empty
request and moves the ResourceReqeust creation from a RequestSpec to a
static factory method. This is a preparation to introduce another
factory method later that will generate a ResourceRequest from a single
ResourceGroup instead of a full RequestSpec.

Blueprint: support-interface-attach-with-qos-ports

Change-Id: Idd58298a6b01775f962b9bf0a0835f762c8e0ed2