Test numa and vcpu topologies bug: #1910466
This change reproduces bug #1910466
When hw:cpu_max_[sockets|cores|threads] is configured
in addition to an explict numa topologies and cpu pinning
nova is currently incapable of generating the correct
virtual CPU topology resulting in an index out of range
error as we attempt to retrieve the first topology
from an empty list.

This change reproduces the error via a new functional
test.

Related-Bug: #1910466
Change-Id: I333b3d85deed971678141307dd06545e308cf989